third_party/blink/web_tests/external/wpt/html/browsers/sandboxing/sandbox-inherited-from-required-csp-expected.txt

This is a testharness.js-based test.
[FAIL] initial empty document
  assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
Harness: the test ran to completion.
